requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-render-partial==0.4
django-simple-captcha==0.5.13
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
Pillow==8.0.1
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
